refactor: content of src/lib of staking should go to bzx-common/src/lib
