+++ packages/bzx-common/src/lib/hashUtils.ts
