+++ packages/bzx-common/src/lib/numberUtils.ts
