+++ packages/staking-dashboard/src/App/Staking/Governance/index.ts
@@ -1,4 +1,4 @@
-import withVM from 'app-lib/withVM'
+import withVM from 'bzx-common/src/lib/withVM'
 import Governance from './Governance'
 import GovernanceVM from './GovernanceVM'
 export default withVM(Governance, GovernanceVM)
