+++ packages/staking-dashboard/src/App/Staking/Rewards/index.ts
@@ -1,4 +1,4 @@
-import withVM from 'app-lib/withVM'
+import withVM from 'bzx-common/src/lib/withVM'
 import Rewards from './Rewards'
 import RewardsVM from './RewardsVM'
 export default withVM(Rewards, RewardsVM)
