+++ packages/staking-dashboard/src/App/Staking/StakingForm/StakingFormVM.ts
@@ -1,5 +1,5 @@
 import { BigNumber } from '@0x/utils'
-import stakingUtils from 'app-lib/stakingUtils'
+import stakingUtils from 'bzx-common/src/lib/stakingUtils'
 import * as mobx from 'mobx'
 import appConfig from 'src/config/appConfig'
 import { stakeableToken } from 'src/domain/stakingTypes'
