+++ packages/staking-dashboard/src/App/Staking/StakingForm/index.ts
@@ -1,4 +1,4 @@
-import withVM from 'app-lib/withVM'
+import withVM from 'bzx-common/src/lib/withVM'
 import StakingForm from './StakingForm'
 import StakingFormVM from './StakingFormVM'
 export default withVM(StakingForm, StakingFormVM)
