+++ packages/staking-dashboard/src/App/index.ts
@@ -1,4 +1,4 @@
-import withVM from 'app-lib/withVM'
+import withVM from 'bzx-common/src/lib/withVM'
 import App from './App'
 import AppVM from './AppVM'
 export default withVM(App, AppVM)
