+++ packages/staking-dashboard/src/stores/RootStore.ts
@@ -5,7 +5,7 @@ import TransactionStore from './TransactionStore'
 import UIStore from './UIStore'
 import Web3Connection from './Web3Connection'
 import appConfig from 'src/config/appConfig'
-import errorUtils from 'app-lib/errorUtils'
+import errorUtils from 'bzx-common/src/lib/errorUtils'
 import GovernanceStore from './GovernanceStore'
 
 export default class RootStore {
