+++ packages/staking-dashboard/src/stores/StakingStore/SpendingAllowance.ts
@@ -1,5 +1,5 @@
 import { BigNumber } from '@0x/utils'
-import errorUtils from 'app-lib/errorUtils'
+import errorUtils from 'bzx-common/src/lib/errorUtils'
 import * as mobx from 'mobx'
 import Asset from 'src/domain/Asset'
 import { StakingProvider } from 'src/services/StakingProvider'
