+++ packages/torque/package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "@0x/subproviders": "^6.3.2",
     "@0x/utils": "^6.2.0",
+    "@types/jest": "^26.0.20",
     "@types/rc-slider": "^8.6.5",
     "events": "^3.0.0",
     "moment": "^2.22.1",
@@ -45,10 +46,10 @@
     "@typescript-eslint/parser": "^4.14.0",
     "@walletconnect/types": "^1.0.0-beta.22",
     "cross-env": "^7.0.3",
+    "git-commit-id": "2.0.1",
     "prettier": "^2.2.1",
     "react-scripts": "^4.0.1",
     "tslint": "^6.1.3",
-    "git-commit-id": "2.0.1",
     "tslint-config-prettier": "^1.18.0",
     "tslint-config-standard": "^9.0.0",
     "tslint-react": "^5.0.0",
